twitter OAuthの実装

+++ database/migrations/2019_02_17_075144_questions.php
@@ -15,7 +15,7 @@ class Questions extends Migration
     {
         Schema::create('questions', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('auther_id')->nullable();  // OAuth等で連携する際に利用
+            $table->bigInteger('auther_id')->nullable();  // OAuth等で連携する際に利用
             $table->string('auther_name');
             $table->string('question_title');
             $table->string('question_detail');
